Revert "Внутренние вычисления переведены в формат decimal для большей точности. Также добавлено округление вверх до тысячных при рассчёте потоков, чтобы нивелировать неточности, возникающие при работе с нечётными дробями."
This reverts commit 8d0eae951cba5e6449c0c15f2d13b24f7bb54ea9.

+++ Model/Elements/ResourceOwerflow.cs
@@ -5,10 +5,10 @@ namespace SatisfactoryProductionManager.Model.Elements
     public class ResourceOverflow
     {
         public string Resource { get; set; }
-        public decimal CountPerMinute { get; set; }
+        public double CountPerMinute { get; set; }
 
 
-        public ResourceOverflow(string resource, decimal count)
+        public ResourceOverflow(string resource, double count)
         {
             Resource = resource;
             CountPerMinute = count;
